+++ sbnanaobj/StandardRecord/SRParticleTruthDLP.h
@@ -42,7 +42,7 @@ namespace caf
         double csda_ke_per_pid[6];                     //!< CSDA kinetic energy per PID.
         float depositions_adapt_q_sum;                 //!< Total tagged (reco non-ghost) charge deposited [ADC].
         float depositions_adapt_sum;                   //!< Total tagged (reco non-ghost) energy deposited [MeV].
-        int64_t depositions_g4_sum;                    //!< Total energy deposited energy at the G4 level [MeV].
+        float depositions_g4_sum;                      //!< Total energy deposited energy at the G4 level [MeV].
         float depositions_q_sum;                       //!< Total tagged (true non-ghost) charge deposited [ADC].
         float depositions_sum;                         //!< Total tagged (true non-ghost) energy deposited [MeV].
         double distance_travel;                        //!< Distance traveled by the neutrino from production to the interaction.
